Separa configuración de firebase y las funciones que se están haciendo de firebase
