+++ src/firebase/index.js
@@ -1,22 +1,11 @@
 /* eslint-disable max-len */
 /* eslint-disable import/no-extraneous-dependencies */
-import { initializeApp } from 'firebase/app';
 import { getFirestore, setDoc, doc } from 'firebase/firestore';
 import {
   getAuth, createUserWithEmailAndPassword, updateProfile, signInWithEmailAndPassword, GoogleAuthProvider, signInWithPopup,
 } from 'firebase/auth';
+import { app } from './init.js';
 
-const firebaseConfig = {
-  apiKey: 'AIzaSyCQM1c-S_xesXYnQCp8NQXpWSk_yvfAZE0',
-  authDomain: 'rockbook-a84e0.firebaseapp.com',
-  projectId: 'rockbook-a84e0',
-  storageBucket: 'rockbook-a84e0.appspot.com',
-  messagingSenderId: '259755500176',
-  appId: '1:259755500176:web:1eaa6528cfd8091f4fae15',
-  measurementId: 'G-17VRS4G1DD',
-};
-
-export const app = initializeApp(firebaseConfig);
 export const auth = getAuth(app);
 export const db = getFirestore(app);
