+++ src/firebase/init.js
@@ -1,0 +1,15 @@
+/* eslint-disable max-len */
+/* eslint-disable import/no-extraneous-dependencies */
+import { initializeApp } from 'firebase/app';
+
+const firebaseConfig = {
+  apiKey: 'AIzaSyCQM1c-S_xesXYnQCp8NQXpWSk_yvfAZE0',
+  authDomain: 'rockbook-a84e0.firebaseapp.com',
+  projectId: 'rockbook-a84e0',
+  storageBucket: 'rockbook-a84e0.appspot.com',
+  messagingSenderId: '259755500176',
+  appId: '1:259755500176:web:1eaa6528cfd8091f4fae15',
+  measurementId: 'G-17VRS4G1DD',
+};
+
+export const app = initializeApp(firebaseConfig);
